+++ src/main/java/com/picimako/justkitting/ServiceLevelDecider.java
@@ -2,13 +2,9 @@
 
 package com.picimako.justkitting;
 
-import static com.picimako.justkitting.PlatformNames.SERVICE_ANNOTATION;
-import static java.util.stream.Collectors.toList;
-
 import com.intellij.codeInsight.AnnotationUtil;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiClass;
-import com.intellij.psi.PsiManager;
 import com.intellij.psi.PsiNameIdentifierOwner;
 import com.intellij.psi.PsiReferenceExpression;
 import com.picimako.justkitting.resources.JustKittingBundle;
@@ -17,16 +13,16 @@ import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.kotlin.asJava.elements.KtLightPsiLiteral;
+import org.jetbrains.kotlin.psi.KtAnnotationEntry;
 import org.jetbrains.kotlin.psi.KtClass;
-import org.jetbrains.uast.UClass;
-import org.jetbrains.uast.UastFacade;
-import org.jetbrains.uast.kotlin.KotlinUQualifiedReferenceExpression;
-import org.jetbrains.uast.kotlin.KotlinUVarargExpression;
+import org.jetbrains.kotlin.psi.ValueArgument;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.regex.Pattern;
+
+import static com.picimako.justkitting.PlatformNames.SERVICE_ANNOTATION;
 
 /**
  * Utility for determining the service level of classes that are annotated with {@link com.intellij.openapi.components.Service}.
@@ -66,53 +62,36 @@ public final class ServiceLevelDecider {
         } else {
             //NOTE: existing constructor and parameter injection are not taken into account in the determination process, for now
             String className = targetClass.getName();
-            if (PROJECT_SERVICE_CLASS_NAME_PATTERN.matcher(className).matches()) {
+            if (PROJECT_SERVICE_CLASS_NAME_PATTERN.matcher(className).matches())
                 level = ServiceLevel.PROJECT;
-            } else if (APP_SERVICE_CLASS_NAME_PATTERN.matcher(className).matches()) {
+            else if (APP_SERVICE_CLASS_NAME_PATTERN.matcher(className).matches())
                 level = ServiceLevel.APP;
-            }
         }
         return level;
     }
 
     /**
      * Returns the {@link com.intellij.openapi.components.Service.Level} values specified in the target class' {@code @Service} annotation.
+     * <p>
+     * These values are then handled in {@link #convertToServiceLevelNames(List, Project)}.
      */
-    private static <T extends PsiNameIdentifierOwner> @Nullable List<? extends Object> getSpecifiedServiceLevels(@NotNull T targetClass) {
-        if (targetClass instanceof PsiClass) {
-            var serviceAnnotation = ((PsiClass) targetClass).getAnnotation(SERVICE_ANNOTATION);
-            if (serviceAnnotation == null) return null;
-
-            var serviceValueAttr = serviceAnnotation.findAttributeValue("value");
-            return AnnotationUtil.arrayAttributeValues(serviceValueAttr);
+    private static <T extends PsiNameIdentifierOwner> @Nullable List<?> getSpecifiedServiceLevels(@NotNull T targetClass) {
+        if (targetClass instanceof PsiClass javaServiceClass) {
+            var serviceAnnotation = javaServiceClass.getAnnotation(SERVICE_ANNOTATION);
+            return serviceAnnotation != null
+                ? AnnotationUtil.arrayAttributeValues(serviceAnnotation.findAttributeValue("value"))
+                : null;
         }
 
-        /*
-         * Since I didn't find a proper way to fetch the annotation from a KtClass by the annotation's fully qualified name,
-         * and I didn't find a way to retrieve the fully qualified name of a KtAnnotation or KtAnnotationEntry,
-         * it spiraled into this UAST-based implementation.
-         *
-         * It works.
-         */
-        if (targetClass instanceof KtClass) {
-            var targetCls = UastFacade.INSTANCE.convertElementWithParent(targetClass, new Class[]{UClass.class});
-            if (targetCls instanceof UClass) {
-                var serviceAnnotation = ((UClass) targetCls).findAnnotation(SERVICE_ANNOTATION);
-                if (serviceAnnotation == null) return null;
-
-                var serviceValueAttr = serviceAnnotation.findAttributeValue("value");
-                if (serviceValueAttr instanceof KotlinUVarargExpression) {
-                    return ((KotlinUVarargExpression) serviceValueAttr).getValueArguments()
-                        .stream()
-                        .filter(KotlinUQualifiedReferenceExpression.class::isInstance)
-                        .map(KotlinUQualifiedReferenceExpression.class::cast)
-                        .collect(toList());
-                } else if (serviceValueAttr instanceof KotlinUQualifiedReferenceExpression) {
-                    return Collections.singletonList(serviceValueAttr);
-                }
-            }
-        }
-        return null;
+        return targetClass instanceof KtClass kotlinServiceClass
+            ? kotlinServiceClass.getAnnotationEntries()
+            .stream()
+            //Find the @Service annotation's entry by its name
+            .filter(entry -> entry.getShortName() != null && "Service".equals(entry.getShortName().asString()))
+            .findFirst()
+            .map(KtAnnotationEntry::getValueArguments)
+            .orElse(null)
+            : null;
     }
 
     /**
@@ -126,26 +105,33 @@ public final class ServiceLevelDecider {
         var cache = PlatformPsiCache.getInstance(project);
         return serviceLevels.stream()
             .map(expression -> {
-                if (expression instanceof PsiReferenceExpression) {
-                    var levelRef = (PsiReferenceExpression) expression;
-                    return levelRef.isReferenceTo(cache.getServiceLevelProject()) || levelRef.isReferenceTo(cache.getServiceLevelApp()) ? levelRef.getReferenceName() : null;
-                } else if (expression instanceof KotlinUQualifiedReferenceExpression) {
-                    var levelRef = (KotlinUQualifiedReferenceExpression) expression;
-                    var levelEnumConst = levelRef.resolve();
-
-                    var psiManager = PsiManager.getInstance(project);
-                    return psiManager.areElementsEquivalent(levelEnumConst, cache.getServiceLevelProject()) || psiManager.areElementsEquivalent(levelEnumConst, cache.getServiceLevelApp())
-                        ? levelRef.getResolvedName()
+                //Handles Java annotation values
+                if (expression instanceof PsiReferenceExpression levelRef)
+                    return levelRef.isReferenceTo(cache.getServiceLevelProject()) || levelRef.isReferenceTo(cache.getServiceLevelApp())
+                        ? levelRef.getReferenceName()
                         : null;
-                } else if (expression instanceof KtLightPsiLiteral) {
-                    var levelRefValue = ((KtLightPsiLiteral) expression).getValue();
-                    return levelRefValue instanceof Pair ? ((Pair) levelRefValue).getSecond().toString() : null;
+
+                //Handles Kotlin annotation values
+                if (expression instanceof ValueArgument levelArg) {
+                    String argumentText = levelArg.asElement().getText();
+
+                    /*
+                     * Returns the service level type based on if the annotation value ends with the proper string.
+                     * This is a very simplified logic that handles cases when the service level is specified with
+                     * or without a qualifier.
+                     */
+                    if (argumentText.endsWith(PROJECT)) return PROJECT;
+                    if (argumentText.endsWith(APP)) return APP;
+                }
+                //Handles Kotlin cases at least in integration tests, but might occur in the wild too.
+                else if (expression instanceof KtLightPsiLiteral levelRef) {
+                    return levelRef.getValue() instanceof Pair levelRefValue ? levelRefValue.getSecond().toString() : null;
                 }
                 return null;
             })
             .filter(Objects::nonNull)
             .distinct()
-            .collect(toList());
+            .toList();
     }
 
     private ServiceLevelDecider() {